Move default tokenizer to Core
