+++ service/Core/AI/DefaultGPTTokenizer.cs
@@ -1,13 +1,12 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
 using System.Collections.Generic;
+using System.Linq;
 using Microsoft.ML.Tokenizers;
 
-#pragma warning disable IDE0130 // reduce number of "using" statements
-// ReSharper disable once CheckNamespace
-namespace Microsoft.KernelMemory.AI.OpenAI;
+namespace Microsoft.KernelMemory.AI;
 
-public static class DefaultGPTTokenizer
+public class DefaultGPTTokenizer : ITextTokenizer
 {
     private static readonly Tokenizer s_tokenizer = TiktokenTokenizer.CreateForModel(
         "gpt-4", new Dictionary<string, int> { { "<|im_start|>", 100264 }, { "<|im_end|>", 100265 } });
@@ -15,5 +14,15 @@ public static class DefaultGPTTokenizer
     public static int StaticCountTokens(string text)
     {
         return s_tokenizer.CountTokens(text);
+    }
+
+    public int CountTokens(string text)
+    {
+        return s_tokenizer.CountTokens(text);
+    }
+
+    public IReadOnlyList<string> GetTokens(string text)
+    {
+        return s_tokenizer.EncodeToTokens(text, out string? _).Select(t => t.Value).ToList();
     }
 }
